fix unwanted promotion in dense and pytree
(oks would be double precition even when params and samples were single 
precision)

+++ netket/optimizer/qgt/qgt_jacobian_dense_logic.py
@@ -287,7 +287,7 @@ def prepare_centered_oks(
     )
 
     n_samp = samples.shape[0] * mpi.n_nodes
-    centered_oks = subtract_mean(jacobians, axis=0) / np.sqrt(n_samp)
+    centered_oks = subtract_mean(jacobians, axis=0) / np.sqrt(n_samp, dtype=jacobians.dtype)
 
     centered_oks = centered_oks.reshape(-1, centered_oks.shape[-1])
